feat:#86 sidebar client컴포넌트 추가
